third_party/WebKit/LayoutTests/fast/encoding/api/ascii-supersets.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
<script>

description("Supersets of ASCII decode ASCII correctly");

// Encodings that have escape codes in 0x00-0x7F
var escape_codes = {
    "hz-gb-2312": [ 0x7E ],
    "iso-2022-jp": [ 0x1B ],
    "iso-2022-kr": [ 0x0E, 0x0F, 0x1B ]
};

encodings_table.forEach(function(section) {
    section.encodings.forEach(function(encoding) {
        if (encoding.name === "replacement")
            return;
        if (utf_encodings.indexOf(encoding.name) !== -1)
            return;

        string = '';
        decoded = null;
        bytes = [];
        for (var i = 0; i < 128; ++i) {
            if (encoding.name in escape_codes && escape_codes[encoding.name].indexOf(i) !== -1)
                continue;
            string += String.fromCharCode(i);
            bytes.push(i);
        }

        decoder = null;
        evalAndLog("decoder = new TextDecoder(" + JSON.stringify(encoding.name) + ")");
        if (decoder) {
            evalAndLog("decoded = decoder.decode(new Uint8Array(bytes))");
            // encodeURIComponent ensures output is printable
            shouldBe("encodeURIComponent(string)", "encodeURIComponent(decoded)");
        }
    });
});


</script>
<script src="../../js/resources/js-test-post.js"></script>
