third_party/WebKit/LayoutTests/fast/encoding/api/ascii-supersets-expected.txt

Supersets of ASCII decode ASCII correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


decoder = new TextDecoder("ibm866")
FAIL decoder = new TextDecoder("ibm866") threw exception TypeError: Type error
decoder = new TextDecoder("iso-8859-2")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-3")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-4")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-5")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-6")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-7")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-8")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-8-i")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-10")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-13")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-14")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-15")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-8859-16")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("koi8-r")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("koi8-u")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("macintosh")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-874")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1250")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1251")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1252")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1253")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1254")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1255")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1256")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1257")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("windows-1258")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("x-mac-cyrillic")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("gbk")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("gb18030")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("hz-gb-2312")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("big5")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("euc-jp")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-2022-jp")
decoded = decoder.decode(new Uint8Array(bytes))
FAIL encodeURIComponent(string) should be %00%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%EF%BF%BD%EF%BF%BD%10%11%12%13%14%15%16%17%18%19%1A%1C%1D%1E%1F%20!%22%23%24%25%26'()*%2B%2C-.%2F0123456789%3A%3B%3C%3D%3E%3F%40ABCDEFGHIJKLMNOPQRSTUVWXYZ%5B%5C%5D%5E_%60abcdefghijklmnopqrstuvwxyz%7B%7C%7D~%7F. Was %00%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%0E%0F%10%11%12%13%14%15%16%17%18%19%1A%1C%1D%1E%1F%20!%22%23%24%25%26'()*%2B%2C-.%2F0123456789%3A%3B%3C%3D%3E%3F%40ABCDEFGHIJKLMNOPQRSTUVWXYZ%5B%5C%5D%5E_%60abcdefghijklmnopqrstuvwxyz%7B%7C%7D~%7F.
decoder = new TextDecoder("shift_jis")
decoded = decoder.decode(new Uint8Array(bytes))
FAIL encodeURIComponent(string) should be %00%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%0E%0F%10%11%12%13%14%15%16%17%18%19%1C%1B%7F%1D%1E%1F%20!%22%23%24%25%26'()*%2B%2C-.%2F0123456789%3A%3B%3C%3D%3E%3F%40ABCDEFGHIJKLMNOPQRSTUVWXYZ%5B%5C%5D%5E_%60abcdefghijklmnopqrstuvwxyz%7B%7C%7D~%1A. Was %00%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%0E%0F%10%11%12%13%14%15%16%17%18%19%1A%1B%1C%1D%1E%1F%20!%22%23%24%25%26'()*%2B%2C-.%2F0123456789%3A%3B%3C%3D%3E%3F%40ABCDEFGHIJKLMNOPQRSTUVWXYZ%5B%5C%5D%5E_%60abcdefghijklmnopqrstuvwxyz%7B%7C%7D~%7F.
decoder = new TextDecoder("euc-kr")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("iso-2022-kr")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
decoder = new TextDecoder("x-user-defined")
decoded = decoder.decode(new Uint8Array(bytes))
PASS encodeURIComponent(string) is encodeURIComponent(decoded)
PASS successfullyParsed is true

TEST COMPLETE
